renamed new files back to normal names :)
