+++ mobile/src/reducers/app.reducer.ts
@@ -10,9 +10,9 @@ import { userInfoReducer, UserInfoState } from './user/user-info.reducer';
 import { userEpics } from './user/user.reducer';
 import { feedsEpics, feedsReducer, FeedsState } from './feeds/feeds.reducer';
 import { calendarEpics, calendarReducer, CalendarState } from './calendar/calendar.reducer';
-import { SecuredApiClient } from '../auth/secured-api-client2';
+import { SecuredApiClient } from '../auth/secured-api-client';
 import config from '../config';
-import { OAuthProcess } from '../auth/oauth-process-2';
+import { OAuthProcess } from '../auth/oauth-process';
 import { peopleReducer, PeopleState } from './people/people.reducer';
 import { authEpics$, authReducer, AuthState } from './auth/auth.reducer';
 import { refreshEpics } from './refresh/refresh.reducer';
